Created usermgmnt app, and added Templates for Thank you Page, Error Page, Confirm Deletion Page.
