+++ app/emailTemplates/ConfirmUserAccountDelete.html
@@ -24,7 +24,7 @@
 
     
     <p>
-        <a href="{{delete_url}}/v1/api/delete-account/{{uniqueID}}">
+        <a href="{{delete_url}}/v1/user/delete-account/{{uniqueID}}">
             <button type="button" class="button">Delete Account</button>
         </a>
     </p>
