+++ app/urls.py
@@ -42,9 +42,5 @@ urlpatterns = [
     re_path(r'volunteer-profile/', views.VolunteerProfile.as_view(), name='volunteer-profile'), 
     # GET (to fetch vehicle Details), POST (to Add Vehicle Details) and PUT (to Update Vehicle Details)
     re_path(r'volunteer-vehicle/', views.VehicleOperations.as_view(), name='volunteer-vehicle'), 
-    # DELETE (Django Template view to Delete Volunteer Profile)
-    re_path(r'delete-account/(?P<uniqueID>[-\w]*)/', views.DeleteUserAccountView.as_view(), name='delete-account-view'), 
-    # DELETE (Delete Action for the Django Template)
-    re_path(r'delete-account-action/<(?P<uniqueID>[-\w]*)/', views.deleteUserAccountAction, name='delete-account-action'), 
 
 ]+static(settings.MEDIA_URL, documne_root=settings.MEDIA_ROOT)
