+++ usermgmnt/templates/EmailConfirmationForUserDeletion.html
@@ -1,3 +1,4 @@
+{% load static %}
 <!DOCTYPE html>
 <html lang="en">
   <head>
@@ -12,16 +13,18 @@
   </head>
 
   <body>
-    <div><div>
-    <div
-      class="col"
-      style="
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin-top: 150px;
-      "
-    >
+    <div class='container-fluid' style="background-color:rgb(0, 75, 36)">
+      <div class="d-flex flex-row text-white">
+        <div>
+          <img src="{% static 'CH-LOG-MEDIA-STATIC.png' %}">
+        </div>
+        <div class="mx-auto">
+          <h1 >Climate Healers</h1>
+        </div>
+      </div> 
+  </div>
+
+    <div class="col" style=" display: flex; flex-direction: column; align-items:center; margin-top: 100px;">
       <h4 class="typography" direction="horizontal">
         <span
           role="img"
@@ -47,20 +50,18 @@
           </svg>
         </span>
       </h4>
-      <span
-        class="typography"
-        direction="horizontal"
-        style="text-align: left; font-size: 1.8rem; font-weight: bold"
-      >
+
+      <span class="typography" direction="horizontal" style="text-align: left; font-size: 1.8rem; font-weight: bold">
         Account Deletion
       </span>
-      <span class="typography" style="font-size: 1rem">
-        Click on the delete button to confirm the account deletion
+
+      <span class="typography" style="text-align: center; font-size: 1rem">
+        Click on the Delete Account button to confirm your account deletion
       </span>
 
-      <form method="post" action="{% url 'app:delete-account-action' uniqueID %}">
+      <form method="post" action="{% url 'usermgmnt:delete-account-action' uniqueID %}">
         {% csrf_token %}
-        <button type="submit" class="btn btn-primary" style="margin-top: 10px">
+        <button type="submit" class="btn btn-warning bold" style="margin-top: 10px">
           <span>Delete Account</span>
         </button>
       </form>
