+++ usermgmnt/templates/ThankYouPage.html
@@ -1,0 +1,66 @@
+{% load static %}
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Climate Healers</title>
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet">
+    </head>
+
+    <body>
+        <div class='container-fluid' style="background-color:rgb(0, 75, 36)">
+            <div class="d-flex flex-row text-white">
+              <div>
+                <img src="{% static 'CH-LOG-MEDIA-STATIC.png' %}">
+              </div>
+              <div class="mx-auto">
+                <h1 >Climate Healers</h1>
+              </div>
+            </div> 
+        </div>
+
+        <div class="d-flex justify-content-center align-items-center" style="margin-top: 100px;">
+            {% if not success %}
+            <div>
+                <div class="text-center mb-4">
+                    <h1>OOP's Something Went Wrong</h1>
+                </div>
+                <div class="mb-4 text-center">
+                    <svg xmlns="http://www.w3.org/2000/svg" class="text-danger" width="75" height="75" fill="currentColor" class="bi bi-x-circle-fill" viewBox="0 0 16 16">
+                        <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM5.354 4.646a.5.5 0 1 0-.708.708L7.293 8l-2.647 2.646a.5.5 0 0 0 .708.708L8 8.707l2.646 2.647a.5.5 0 0 0 .708-.708L8.707 8l2.647-2.646a.5.5 0 0 0-.708-.708L8 7.293 5.354 4.646z"/>
+                    </svg>                   
+                </div>
+                <div class="text-center">
+                    <h4>Please Try Again !</h4>
+                    <p>You can close the page</p>
+                    <h4>or</h4>
+                    <h4>Contact Climate Healers Team</h4>
+                    <p><a href='https://climatehealers.org/contact-2/contact-us/'>Contact Climate Healers</a></p>
+                </div>
+            </div>
+           {% else %}
+            <div>
+                <div class="text-center mb-4">
+                    <h1>Your Account has been Deleted Successfully</h1>
+                </div>
+                <div class="mb-4 text-center">
+                    <svg xmlns="http://www.w3.org/2000/svg" class="text-success" width="75" height="75" fill="currentColor" class="bi bi-check-circle-fill" viewBox="0 0 16 16">
+                        <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zm-3.97-3.03a.75.75 0 0 0-1.08.022L7.477 9.417 5.384 7.323a.75.75 0 0 0-1.06 1.06L6.97 11.03a.75.75 0 0 0 1.079-.02l3.992-4.99a.75.75 0 0 0-.01-1.05z" />
+                    </svg>                  
+                </div>
+                <div class="text-center">
+                    <h4>Thank You !</h4>
+                    <p>You can close the page</p>
+                    <p>or</p>
+                    <p>Be a part of the Change <a href='https://youtu.be/b8QRVHJMXOw'>View Climate Healers</a></p>
+                </div>
+            </div>
+            {% endif %}
+        </div>
+    </body>
+
+</html>
+
+
